wip command structure
Changes: wip work on reverse engineering the command structure

Details:

Two FTDI USB -> UART TTL converters are used to be a man in the middle
between the panel and radio allowing me to inject serial data.
This isn't super ideal as the radio's own responses will interfere and
there probably is collision, still it seems pretty safe to do.